move shared constants out of __init__.py
